Move some configure checks to runtime
We were allowing remote clang preprocessing entirely on the basis of
whether the clang Icecream was built with would accept the
-frewrite-includes option. However, there was a subtle bug in the
configure check that meant the option was never actually detected
properly, which essentially always disallowed remote preprocessing with
clang unless CLANG_REMOTE_CPP=1 was set.

Additionally, remote preprocessing with clang should not be conditioned
on -frewrite-includes anyway because that's not the only option which
should matter. We should simply treat clang as an equal to GCC and let
the user decide at runtime whether remote preprocessing should be
disallowed on the basis of what they know about their toolchain.

Finally, the entire basis of these compile checks was that GCC might be
misdetected on Apple platforms due to its path. A simpler and more
accurate check would be looking for GCC in a system path and overriding
the detection to Clang on Apple platforms only, which means we don't
need these compile checks at all.